# Template rendering optimization