resolves #7449 

### Description

Reduces the repetitive rendering described in the issue by caching results.

Local before on internal analytics:

```
dbt --no-version-check --no-partial-parse parse  20.67s user 0.51s system 98% cpu 21.506 total
```

And after:

```
dbt --no-version-check --no-partial-parse parse  18.71s user 0.49s system 98% cpu 19.465 total
```

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-core/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] I have run this code in development and it appears to resolve the stated issue
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have [opened an issue to add/update docs](https://github.com/dbt-labs/docs.getdbt.com/issues/new/choose), or docs changes are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-core/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)