@dbeatty10 2023-04-24
Love the sweet sounds of "12% faster" @peterallenwebb 🎵 !

Curious: is the caching approach a form of [memoization](https://en.wikipedia.org/wiki/Memoization)?