@gshank 2023-04-24
Is this one of those global file things that we need to reset for new commands?  Some of the values in the strings that are rendered could change from command to command, such as variables, etc. So it should be reset at some point.