@peterallenwebb 2023-04-24
@dbeatty10 Yes, it is effectively memoizing this function.