@peterallenwebb 2023-04-24
@gshank I think that is a good point about clearing the cache between commands, and it makes me wonder if we are doing this for other caches (or should). I wish I felt like I could trust the comment, because if I did we could just call python's ast.literal_eval and return the result (or a the original string if it throws).