@peterallenwebb 2023-05-30
@jtcohen6 I'd been putting it off because i have a more comprehensive solution in mind, but I don't think there is any reason not to enjoy this little speedup in the meantime.